requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.2
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 
PyQt5==5.15.9
requests==2.31.0
email-validator==2.2.0
matplotlib==3.9.4
python-dotenv==1.0.0
bcrypt==4.0.1
